+++ collections/store.go
@@ -12,6 +12,18 @@ import (
 
 const filePerm = 0755
 
+func Exists(filePath string) bool {
+	_, err := os.Stat(filePath)
+	if err == nil {
+		return true
+	}
+	if os.IsNotExist(err) {
+		return false
+	}
+	return true
+}
+
+// Save a collection
 func Save(c *Collection) error {
 	if Exists(c.Metadata.CollectionRoot) {
 		return newErr("cannot create collection as a collection with this name already exists", nil, log.Data{"name": c.Name})
@@ -28,26 +40,7 @@ func Save(c *Collection) error {
 	return nil
 }
 
-func Exists(filePath string) bool {
-	_, err := os.Stat(filePath)
-	if err == nil {
-		return true
-	}
-	if os.IsNotExist(err) {
-		return false
-	}
-	return true
-}
-
-func IsMoveBlocked(relURI string, cols []*Collection) error {
-	for _, c := range cols {
-		if c.Contains(relURI) {
-			return newErr("cannot complete move as file containing the target uri is in another collections", nil, log.Data{"file": relURI, "collection": c.Name})
-		}
-	}
-	return nil
-}
-
+// Delete a collection
 func Delete(rootPath string, name string) error {
 	target := path.Join(rootPath, name)
 	if !Exists(target) {
@@ -58,29 +51,8 @@ func Delete(rootPath string, name string) error {
 	return os.RemoveAll(target)
 }
 
-func LoadCollections(collectionsRoot string) ([]*Collection, error) {
-	log.Event(nil, "loading existing collections")
-	collectionFiles, err := ioutil.ReadDir(collectionsRoot)
-	if err != nil {
-		return nil, newErr("failed to read collections dir", err, nil)
-	}
-
-	collections := make([]*Collection, 0)
-	for _, f := range collectionFiles {
-		if f.IsDir() {
-			c, err := LoadCollection(collectionsRoot, f.Name())
-			if err != nil {
-				return nil, newErr("failed to load collection", err, log.Data{"collectionName": f.Name()})
-			}
-			if c != nil {
-				collections = append(collections, c)
-			}
-		}
-	}
-	return collections, nil
-}
-
-func LoadCollection(collectionsRoot string, name string) (*Collection, error) {
+// Get a collection by collection.description.name
+func GetCollection(collectionsRoot string, name string) (*Collection, error) {
 	metadata := NewMetadata(collectionsRoot, name)
 	if !Exists(metadata.CollectionJSON) {
 		log.Event(nil, "no collection json file exists for collection", log.Data{"collection": name})
@@ -98,6 +70,50 @@ func LoadCollection(collectionsRoot string, name string) (*Collection, error) {
 	}
 	col.Metadata = metadata
 	return &col, nil
+}
+
+// Get all collections.
+func GetCollections(collectionsRoot string) (*Collections, error) {
+	log.Event(nil, "loading existing collections")
+	collectionFiles, err := ioutil.ReadDir(collectionsRoot)
+	if err != nil {
+		return nil, newErr("failed to read collections dir", err, nil)
+	}
+
+	collections := &Collections{Collections: make([]*Collection, 0)}
+
+	for _, f := range collectionFiles {
+		if f.IsDir() {
+			c, err := GetCollection(collectionsRoot, f.Name())
+			if err != nil {
+				return nil, newErr("failed to load collection", err, log.Data{"collectionName": f.Name()})
+			}
+			collections.Add(c)
+		}
+	}
+	return collections, nil
+}
+
+func IsMoveBlocked(relURI string, cols *Collections, moveCollection *Collection) error {
+	for _, c := range cols.Collections {
+		if c.Name == moveCollection.Name {
+			// skip this one.
+			continue
+		}
+		if c.Contains(relURI) {
+			return newErr("cannot complete move as file containing the target uri is in another collections", nil, log.Data{"file": relURI, "collection": c.Name})
+		}
+	}
+	return nil
+}
+
+func writeContent(uri string, fileBytes []byte) error {
+	dirs, _ := filepath.Split(uri)
+
+	if err := os.MkdirAll(dirs, filePerm); err != nil {
+		return err
+	}
+	return ioutil.WriteFile(uri, fileBytes, filePerm)
 }
 
 func WriteFileToCollection(c *Collection, relPath string, fileBytes []byte) error {
